修改线程数调节的样式

+++ bridge/lib/call_rust/app.dart
@@ -8,7 +8,7 @@ import 'package:flutter_rust_bridge/flutter_rust_bridge_for_generated.dart';
 
 // Rust type: RustOpaqueMoi<flutter_rust_bridge::for_generated::RustAutoOpaqueInner<TaskConfig>>
 abstract class TaskConfig implements RustOpaqueInterface {
-  BigInt get thread;
+  int get thread;
 
   String get tidyFolder;
 }
